refactor: update imports to use _utils module instead of _matrix, resolve more type checking issues
