+++ src/alphapepttools/_utils.py
@@ -6,10 +6,9 @@ The pipeline only ever handles dense numpy arrays, so the union is narrowed in
 one place here rather than at every call site.
 """
 
-from typing import cast
-
 import anndata as ad
 import numpy as np
+from scipy import sparse
 
 
 def get_matrix(adata: ad.AnnData, layer: str | None = None) -> np.ndarray:
@@ -28,4 +27,6 @@ def get_matrix(adata: ad.AnnData, layer: str | None = None) -> np.ndarray:
         The requested matrix.
     """
     data = adata.X if layer is None else adata.layers[layer]
-    return cast("np.ndarray", data)
+    if sparse.issparse(data):  # this is to cover anndata backends returning sparse data, e.g. csr_matrix
+        data = data.toarray()
+    return np.asarray(data)
